+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/groups/groupDetail/GroupDetailScreen.kt
@@ -441,7 +441,8 @@ fun GroupTabRow(
                 modifier = indicatorModifier
             )
         },
-        divider = {}
+        divider = {},
+        minTabWidth = 0.dp
     ) {
         val coroutineScope = rememberCoroutineScope()